requirements.txt
anthropic>=0.25.0
requests>=2.31.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
sqlalchemy>=2.0.0
tenacity>=8.2.3
rich>=13.7.0
pytest>=8.0.0
